+++ crates/batbox-file-dialog/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "batbox-file-dialog"
+authors.workspace = true
+categories.workspace = true
+description.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+futures.workspace = true
+anyhow.workspace = true
+batbox-file.workspace = true
+
+[target.'cfg(not(target_arch="wasm32"))'.dependencies]
+tinyfiledialogs.workspace = true
+
+[target.'cfg(target_arch="wasm32")'.dependencies]
+web-sys = { workspace = true, features = [
+    "HtmlAnchorElement",
+    "HtmlInputElement",
+    "Document",
+    "Url",
+    "Blob",
+    "File",
+    "FileList",
+] }
+js-sys.workspace = true
+wasm-bindgen.workspace = true
