+++ crates/batbox-file/Cargo.toml
@@ -1,0 +1,39 @@
+[package]
+name = "batbox-file"
+authors.workspace = true
+categories.workspace = true
+description.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+version.workspace = true
+
+[dependencies]
+serde.workspace = true
+serde_json.workspace = true
+toml.workspace = true
+ron.workspace = true
+futures.workspace = true
+anyhow.workspace = true
+log.workspace = true
+
+[target.'cfg(not(target_arch="wasm32"))'.dependencies]
+async-std.workspace = true
+reqwest.workspace = true
+async-compat.workspace = true
+url.workspace = true
+
+[target.'cfg(target_arch="wasm32")'.dependencies]
+web-sys = { workspace = true, features = [
+    "ReadableStream",
+    "Response",
+    "Window",
+] }
+wasm-streams.workspace = true
+wasm-bindgen.workspace = true
+wasm-bindgen-futures.workspace = true
+js-sys.workspace = true
+http.workspace = true
