+++ crates/batbox/Cargo.toml
@@ -21,6 +21,8 @@ batbox-color.workspace = true
 batbox-approx.workspace = true
 batbox-tuple-macros.workspace = true
 batbox-time.workspace = true
+batbox-file.workspace = true
+batbox-file-dialog.workspace = true
 
 batbox-derive.workspace = true
