+++ crates/batbox/src/lib.rs
@@ -42,6 +42,8 @@ pub mod crates {
 
     pub use ::batbox_approx as approx;
     pub use ::batbox_color as color;
+    pub use ::batbox_file as file;
+    pub use ::batbox_file_dialog as file_dialog;
     pub use ::batbox_i18n as i18n;
     pub use ::batbox_num as num;
     pub use ::batbox_time as time;
@@ -58,7 +60,6 @@ pub use batbox_derive::*;
 pub mod cmp;
 pub mod collection;
 pub mod diff;
-pub mod file;
 pub mod geom;
 pub mod logger;
 pub mod preferences;
